Tests/LibWeb/Text/expected/HTML/MessagePort-transfer-twice.txt

first receipt: re-transferring same port
we good
